+++ native/src/seal/batchencoder.cpp
@@ -29,7 +29,8 @@ namespace seal
         }
 
         auto &context_data = *context_->first_context_data();
-        if (context_data.parms().scheme() != scheme_type::BFV)
+        if (context_data.parms().scheme() != scheme_type::BFV &&
+            context_data.parms().scheme() != scheme_type::CKKS_FV)
         {
             throw invalid_argument("unsupported scheme");
         }