+++ native/src/seal/decryptor.cpp
@@ -68,6 +68,7 @@ namespace seal
             return;
 
         case scheme_type::CKKS:
+        case scheme_type::CKKS_FV:
             ckks_decrypt(encrypted, destination, pool_);
             return;
 
@@ -364,7 +365,8 @@ namespace seal
             throw invalid_argument("encrypted is not valid for encryption parameters");
         }
 
-        if (context_->key_context_data()->parms().scheme() != scheme_type::BFV)
+        auto scheme = context_->key_context_data()->parms().scheme();
+        if (scheme != scheme_type::BFV && scheme != scheme_type::CKKS_FV)
         {
             throw logic_error("unsupported scheme");
         }