+++ native/src/seal/evaluator.h
@@ -198,6 +198,21 @@ namespace seal
 
         /**
         Subtracts two ciphertexts. This function computes the difference of encrypted1
+        and encrypted2, and stores the result in encrypted1. This function only works
+        on CKKS_FV scheme, and it allow ciphertexts of different scale.
+
+        @param[in] encrytped1 The ciphertext to subtract from
+        @param[in] encrypted2 The ciphertext to subtract
+        @throws std::invalid_argument if encrypted1 or encrypted2 is not valid for the
+        encryption parameters
+        @throws std::invalid_argument if encrytped1 or encrypted2 are in different
+        NTT forms
+        @throws std::logic_error if result ciphertext is transparent
+        */
+        void sub_ckks_fv_inplace(Ciphertext &encrypted1, const Ciphertext &encrypted2);
+
+        /**
+        Subtracts two ciphertexts. This function computes the difference of encrypted1
         and encrypted2 and stores the result in the destination parameter.
 
         @param[in] encrypted1 The ciphertext to subtract from
@@ -1049,7 +1064,7 @@ namespace seal
         @param[in] steps The number of steps to rotate (negative left, positive right)
         @param[in] galois_keys The Galois keys
         @param[in] pool The MemoryPoolHandle pointing to a valid memory pool
-        @throws std::logic_error if scheme is not scheme_type::BFV
+        @throws std::logic_error if scheme is not scheme_type::BFV or scheme_type::CKKS_FV
         @throws std::logic_error if the encryption parameters do not support batching
         @throws std::invalid_argument if encrypted or galois_keys is not valid for
         the encryption parameters
@@ -1067,7 +1082,8 @@ namespace seal
             int steps, const GaloisKeys &galois_keys,
             MemoryPoolHandle pool = MemoryManager::GetPool())
         {
-            if (context_->key_context_data()->parms().scheme() != scheme_type::BFV)
+            auto scheme = context_->key_context_data()->parms().scheme();
+            if (scheme != scheme_type::BFV && scheme != scheme_type::CKKS_FV)
             {
                 throw std::logic_error("unsupported scheme");
             }
@@ -1089,7 +1105,7 @@ namespace seal
         @param[in] galois_keys The Galois keys
         @param[out] destination The ciphertext to overwrite with the rotated result
         @param[in] pool The MemoryPoolHandle pointing to a valid memory pool
-        @throws std::logic_error if scheme is not scheme_type::BFV
+        @throws std::logic_error if scheme is not scheme_type::BFV or scheme_type::CKKS_FV
         @throws std::logic_error if the encryption parameters do not support batching
         @throws std::invalid_argument if encrypted or galois_keys is not valid for
         the encryption parameters
@@ -1124,7 +1140,7 @@ namespace seal
         @param[in] galois_keys The Galois keys
         @param[out] destination The ciphertext to overwrite with the rotated result
         @param[in] pool The MemoryPoolHandle pointing to a valid memory pool
-        @throws std::logic_error if scheme is not scheme_type::BFV
+        @throws std::logic_error if scheme is not scheme_type::BFV or scheme_type::CKKS_FV
         @throws std::logic_error if the encryption parameters do not support batching
         @throws std::invalid_argument if encrypted or galois_keys is not valid for
         the encryption parameters
@@ -1141,7 +1157,8 @@ namespace seal
             const GaloisKeys &galois_keys,
             MemoryPoolHandle pool = MemoryManager::GetPool())
         {
-            if (context_->key_context_data()->parms().scheme() != scheme_type::BFV)
+            auto scheme = context_->key_context_data()->parms().scheme();
+            if (scheme != scheme_type::BFV && scheme != scheme_type::CKKS_FV)
             {
                 throw std::logic_error("unsupported scheme");
             }
@@ -1161,7 +1178,7 @@ namespace seal
         @param[in] galois_keys The Galois keys
         @param[out] destination The ciphertext to overwrite with the rotated result
         @param[in] pool The MemoryPoolHandle pointing to a valid memory pool
-        @throws std::logic_error if scheme is not scheme_type::BFV
+        @throws std::logic_error if scheme is not scheme_type::BFV or scheme_type::CKKS_FV
         @throws std::logic_error if the encryption parameters do not support batching
         @throws std::invalid_argument if encrypted or galois_keys is not valid for
         the encryption parameters
@@ -1316,6 +1333,39 @@ namespace seal
         {
             destination = encrypted;
             complex_conjugate_inplace(destination, galois_keys, std::move(pool));
+        }
+
+        /**
+        Transforms a FV ciphertext to CKKS ciphertext of the lowest level.
+        Scale of the transformed ciphertext is regarded as its previous scale
+        multiplied with q0/t, which approximates scaling factor in FV scheme.
+
+        @param[in] encrypted_fv The FV ciphertext to transform
+        @throws std::invalid_argument if encrypted_fv is not valid for encryption
+        parameters
+        @throws std::invalid_argument if scheme is not scheme_type::CKKS_FV
+        @throws std::invalid_argument if encrypted_fv is already in NTT form
+        */
+        void transform_to_lowest_ckks_inplace(Ciphertext &encrypted_fv);
+
+        /**
+        Transforms a FV ciphertext to CKKS ciphertext of the lowest level.
+        Scale of the transformed ciphertext is regarded as its previous scale
+        multiplied with q0/t, which approximates scaling factor in FV scheme.
+        The result is stored in the destination parameter.
+
+        @param[in] encrypted_fv The FV ciphertext to transform
+        @param[out] destination The ciphertext to overwrite with the transformed result
+        @throws std::invalid_argument if encrypted_fv is not valid for encryption
+        parameters
+        @throws std::invalid_argument if scheme is not scheme_type::CKKS_FV
+        @throws std::invalid_argument if encrypted_fv is already in NTT form
+        */
+        inline void transform_to_lowest_ckks(const Ciphertext &encrypted_fv,
+            Ciphertext &destination)
+        {
+            destination = encrypted_fv;
+            transform_to_lowest_ckks_inplace(destination);
         }
 
         /**